create `src/validator` dir and move some files into it
