+++ src/validator/rule.test.ts
